refactor: move api.ts to utils directory
